Clarification on README
As mentionned in #71, using the program option for type-checking can lead to `Invalid source file: xxx.ts. Ensure that the files supplied to lint have a .ts, .tsx, .js or .jsx extension.`

Using the base option of gulp.src corrects it, thought mentionning it in README can save some time for users.